Convert `div`-based cover images to lists (incomplete)
Using lists instead of `div`s is more semantically pure and _should_ be more readable when styles are disabled. The problem is that the entire surface of the background image needs to be an active link target, so the `li` is useless for styling. With the `a` used to cover the image area, it needs a child element to style the text overlay. `p` + `span` works well enough and should be legal HTML5, but it doesn't look good in Firefox. I don't like the idea of a `span` with another `span` in it. Hmm.